fix(e2e): resolve major compilation errors and type mismatches
- Fixed float64 to float32 type conversion errors in RAG package
- Removed duplicate Client declarations in porch package
- Fixed struct field mismatches in O1 client interfaces
- Resolved demo compilation conflicts by organizing packages
- Fixed metrics package import and undefined variable errors
- Removed duplicate type declarations in dependencies package
- Fixed duplicate type declarations in optimization package
- Corrected Priority to NetworkPriority type mismatches in tests
- Organized demo files to prevent main function conflicts
- Cleaned up substantial duplicate code (1461 lines removed)

+++ pkg/rag/enhanced_retrieval_service.go
@@ -503,11 +503,11 @@ func (ers *EnhancedRetrievalService) SearchEnhanced(ctx context.Context, request
 
 		HybridSearch: true,
 
-		HybridAlpha: float64(ers.config.DefaultHybridAlpha),
+		HybridAlpha: float32(ers.config.DefaultHybridAlpha),
 
 		UseReranker: false, // We'll do our own reranking
 
-		MinConfidence: float64(request.MinQualityScore),
+		MinConfidence: float32(request.MinQualityScore),
 
 		ExpandQuery: false, // Already done in enhancement step
